BoardService 로그인된 User 검증 기능 추가 Jpa Auditing Annotation 추가

+++ src/main/java/com/sparta/newsfeed_qna/service/BoardService.java
@@ -3,6 +3,7 @@ package com.sparta.newsfeed_qna.service;
 import com.sparta.newsfeed_qna.dto.BoardRequestDto;
 import com.sparta.newsfeed_qna.dto.BoardResponseDto;
 import com.sparta.newsfeed_qna.entity.Board;
+import com.sparta.newsfeed_qna.entity.User;
 import com.sparta.newsfeed_qna.repository.BoardRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Sort;
@@ -18,14 +19,13 @@ public class BoardService {
     private final BoardRepository boardRepository;
 
     // 게시글 작성 API
-    public void createBoard(BoardRequestDto boardRequestDto) {
-        Board board = new Board(boardRequestDto);
-
+    public void createBoard(BoardRequestDto boardRequestDto, User user) {
+        Board board = new Board(boardRequestDto, user);
         boardRepository.save(board);
     }
 
     // 게시글 전체 조회 API
-    public List<BoardResponseDto> getBoard() {
+    public List<BoardResponseDto> getAllBoard() {
         // 작성일 기준 내림차순 Sort.by 사용
         List<BoardResponseDto> boardResponseDtos = boardRepository.findAll(Sort.by(Sort.Direction.DESC, "createdAt")).stream().map(BoardResponseDto::new).toList();
         return boardResponseDtos;
@@ -38,20 +38,27 @@ public class BoardService {
     }
     // 게시글 수정 API
     @Transactional
-    public Long modifyBoard(Long boardId, BoardRequestDto boardRequestDto) {
-        // 해당 게시글이 db에 존재하는지 확인
-        Board board = findBoard(boardId);
-
-        board.update(boardRequestDto);
-        return boardId;
+    public BoardResponseDto modifyBoard(Long boardId, BoardRequestDto boardRequestDto, User user) {
+        // 해당 게시글이 db에 존재하는지 확인, 영속성 컨텍스트 1차 캐시 저장
+        Board board = findBoard(boardId); // boardRepository에서 Board 조회
+        if(board.getUser().getId() == user.getId()){ // 인가받은 user와 repository에 저장된 board의 user가 같은지 확인
+            board.update(boardRequestDto);
+            BoardResponseDto boardResponseDto = new BoardResponseDto(board.getBoardId(), board.getBoardTitle(),
+                    board.getBoardContent(), board.getCreatedAt(), board.getModifiedAt());
+            return boardResponseDto;
+        } else{
+            throw new IllegalArgumentException("해당 게시글의 작성자만 글을 수정할 수 있습니다.");
+        }
     }
 
     // 게시글 삭제 API
-    public Long deleteBoard(Long boardId) {
+    public void deleteBoard(Long boardId, User user) {
         Board board = findBoard(boardId);
+        if(board.getUser().getId() == user.getId()){
+            boardRepository.delete(board);
+        } else {
 
-        boardRepository.delete(board);
-        return boardId;
+        }
     }
 
     // findById를 따로 뺀 메서드
